+++ tests/pos/i6247.scala
@@ -1,0 +1,10 @@
+implicit class NN[T](x:T|Null) {
+  def nn: T = x.asInstanceOf[T]
+}
+
+class Foo {
+  val x1: String|Null = null
+  x1.nn.length
+  val x2: String = x1.nn
+  x1.nn.length
+}
